fix(core): widen MediaSourceInfo/ItemRef bitrate to i64 for negative sentinel
Jellyfin returns Bitrate: -1000 for FLAC tracks whose bitrate could not be
measured, which failed to deserialize into Option<u32> and sank the entire
PlaybackInfo response (and any item listing carrying the sentinel), breaking
playback session tracking for those tracks.

+++ core/src/client.rs
@@ -2632,7 +2632,7 @@ pub struct RawItem {
     #[serde(rename = "Container")]
     pub container: Option<String>,
     #[serde(rename = "Bitrate")]
-    pub bitrate: Option<u32>,
+    pub bitrate: Option<i64>,
     #[serde(rename = "Path")]
     pub path: Option<String>,
     #[serde(rename = "PlaylistItemId")]
